requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
pandas>=1.3.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
mediapipe>=0.8.9
pytesseract>=0.3.8
pyttsx3>=2.90
scikit-learn>=1.0.0